backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
pydantic-settings==2.5.0
python-dotenv==1.0.1
sqlalchemy==2.0.29
psycopg==3.3.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.7
requests==2.32.0
aiohttp==3.10.5
aiofiles==23.2.1
httpx==0.27.2
firebase-admin==6.5.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
super-gradients>=3.3.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.5.0
